some logging cleanup
