+++ .env
@@ -1,5 +1,5 @@
 #### GENERAL CONFIG ####
-# level of logging (1=debug, 2=info, 3=warn, 5=error)
+# level of logging (1=debug, 2=info, 3=warn, 4=error)
 LOG_LEVEL=2
 
 # location of the yaml file that contains type definitions
